000 ARCHIVE/SPAPI_Python_SDK/requirements.txt
requests>=2.25.1
backoff>=2.2.1
boto3>=1.26.0
python-dotenv>=0.19.0
